requirements.txt
feedparser
requests
pytz
msal
tqdm
pyyaml
microsoftgraph-python